+++ .devcontainer/hooks/onCreate.sh
@@ -2,3 +2,5 @@
 set -eux
 
 npm ci
+npm ci --prefix "$HOME/twinejs"
+poetry install -P "$WORKSPACE_DIR/spindler"