+++ .devcontainer/hooks/updateContent.sh
@@ -2,3 +2,5 @@
 set -eux
 
 npm install
+npm install --prefix "$HOME/twinejs"
+poetry install -P "$WORKSPACE_DIR/spindler"