+++ assemblyscript/@gl/api/w2h/debug.ts
@@ -4,5 +4,9 @@
  * @param message The message to log.
  */
 export declare function log(message: string): void;
+export declare function logError(message: string): void;
+export declare function logWarning(message: string): void;
 
 export const _keep_log = log;
+export const _keep_logError = logError;
+export const _keep_logWarning = logWarning;