+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.4.0 - 4/11/25
+
+- TwineJs integration for dialogue authoring
+- Use engine v0.6.0
+- `host.pickup.get(tags)` for querying inventory
+- Unify `displaySign` and `displayInteraction` to `display` with animate arg
+- `sensorEvent` and `tileCollisionEvent` take `initiator` arg
+- 2x WebAssembly compilation speed
+- Update all deps
+- `logError` and `logWarning` now available
+- `host.pickup.get` for checking player's inventory
+- Added `host.time.getSunEvent` for current event ("morning", "dusk", etc)
+- Added template version to compiled WASM file
+
 ## 0.3.3 - 3/26/25
 
 - Faster water movement